eliminación de datos de servicios

+++ app/src/main/java/cl/utem/vote/feeling/manager/VoteManager.java
@@ -94,7 +94,7 @@ public class VoteManager implements Serializable {
             for (Vote vote : votes) {
                 sum += vote.getChoice();
             }
-            avg = sum / total;
+            avg = sum / ((double) total);
             votes.clear();
         }
         return new CountVO(avg, total);
